ENGR00310060 USB:refactor adb function to use ffs
According to AOSP's new usb gadget driver, adb function is removed.
Instead ffs function is for adb. We can make adb work acoording to
the below commit in system/core.

FunctionFS: initial implementation

This is the second version of a patch which demonstrates the possibility
of using adbd (Android Debug Bridge daemon) with a generic
FunctionFS gadget instead of a custom adb usb gadget in the Linux kernel. It
contains changes introduced after Benoit's review - thank you Benoit.

The patch adds a new usb access layer to adbd using FunctionFS. The former
usb access method is still available. The method is chosen at runtime
depending if /dev/usb-ffs/adb/ep0 or /dev/android_adb is accessible.

How to use on the target device:

$ insmod g_ffs.ko idVendor=<vendor ID>
iSerialNumber=<some string>
$ mount -t functionfs adb /dev/usb-ffs/adb -o uid=2000,gid=2000
$ ./adbd

This patch requires a patch to bionic which adds <linux/usb_functionfs.h>
which is an exact copy of the relevant file in the linux kernel.

Signed-off-by: Jianzheng Zhou <Jianzheng.zhou@freescale.com>